Add support for custom request and notification methods
It is relatively common to extend LSP with custom method types that
are specific to a given language so it makes sense for haskell-lsp to
support this.

To accomplish this, this PR introduces the following changes:

1. Remove the `Misc` constructor in favor of a more explicit
   `CustomClientMethod` constructor that does not enforce that methods
   start with $/. While the spec mentions that custom methods should
   start with $/, I’ve seen custom methods that do not start with $/
   and given that haskell-lsp can only control the server side but not
   the client side, e.g., VSCode extensions, I think it makes more
   sense to be relaxed here.

2. Add `ReqCustomClient` and `NotCustomClient` to `FromClientMessage`
   while removing `UnknownFromClientMessage` (any valid JSON-RPC
   message should now be decodable).

3. Add handlers for `ReqCustomClient` and `NotCustomClient`.

4. It adds `ReqCustomServer`, `RspCustomServer` and `NotCustomServer`
   to `FromServerMessage`.

Note that this PR does not add `RspCustomClient`. This is because
responses are always decoded to `BareResponseMessage` with no
restrictions so it would be redundant. Once `BareResponseMessage` is
removed in favor of explicit response types, this constructor and a
corresponding handler should be added.